+++ src/Middleware/PreviewHandler.php
@@ -6,6 +6,7 @@ namespace Primo\Middleware;
 use Laminas\Diactoros\Response\RedirectResponse;
 use Prismic\ApiClient;
 use Prismic\Exception\InvalidPreviewToken;
+use Prismic\Exception\PreviewTokenExpired;
 use Prismic\LinkResolver;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
@@ -46,6 +47,10 @@ final class PreviewHandler implements MiddlewareInterface
             $link = $this->api->previewSession($token);
         } catch (InvalidPreviewToken $error) {
             return $handler->handle($request);
+        } catch (PreviewTokenExpired $expired) {
+            return $handler->handle(
+                $request->withAttribute(PreviewTokenExpired::class, $expired)
+            );
         }
 
         $url = $link ? $this->linkResolver->resolve($link) : $this->defaultUrl;